hunyuan_requirements.txt
accelerate==0.31.0
bitsandbytes==0.42.0
certifi==2022.12.7
charset-normalizer==2.1.1
diffusers==0.17.0
einops==0.8.0
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.23.4
idna==3.4
imagesize==1.4.1
importlib_metadata==7.2.1
Jinja2==3.1.3
lycoris_lora==3.0.0.dev12
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pillow==10.2.0
protobuf==5.27.1
psutil==6.0.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.28.1
safetensors==0.4.3
scipy==1.13.1
sentencepiece==0.2.0
sympy==1.12
timm==1.0.7
tokenizers==0.19.1
toml==0.10.2
torch==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.4
transformers==4.41.2
triton==2.1.0
typing_extensions==4.9.0
urllib3==1.26.13
voluptuous==0.15.0
xformers==0.0.23.post1+cu118
zipp==3.19.2
